Unified around the KL-corrected objective
This commit modifies the existing set of methods to better
capture their connectedness via the KL-corrected objective
function of the VRVM.

An additional variant of the VRVM that uses the Woodbury
identity and the determinant lemma has also been added
to the source tree.

Recent work on the reduced parameterization of the VRVM is
slated for inclusion in this repository as well.