Added IF-SBL, but...
Interestingly enough, the implementation described in the
IF-SBL manuscript *does not run* as advertised. The objective
decreases, but the mean estimate is useless, effectively
remaining identical to A* y.

I'm honestly unsure why this is. I've tested this against
both partial Fourier and normal random matrices, and the
results are the same.

Looking at the published source code, it seems there are
differences from the manuscript, but I also changed my
code to reflect these differences, and the results were
again the same.